Add service class tests for unused fields rule
