+++ scan-command/src/test/resources/testng.xml
@@ -29,7 +29,7 @@ under the License.
             <class name="io.ballerina.scan.internal.IssueImplTest"/>
             <class name="io.ballerina.scan.internal.ReporterImplTest"/>
             <class name="io.ballerina.scan.internal.CoreRuleTest"/>
-            <class name="io.ballerina.scan.internal.StaticCodeAnalyzerTest"/>
+            <class name="io.ballerina.scan.internal.Rule005Test"/>
             <class name="io.ballerina.scan.internal.ProjectAnalyzerTest"/>
             <class name="io.ballerina.scan.utils.ScanUtilsTest"/>
             <class name="io.ballerina.scan.utils.DiagnosticLogTest"/>
